updates the migrations with exists check

+++ resources/Migrations/2017_12_06_160000_CreateULLogsTable.php
@@ -13,22 +13,24 @@ class CreateULLogsTable extends Migration
      */
     public function up()
     {
-        Schema::connection($this->connection)->create('logs', function (Blueprint $table) {
-            $table->bigIncrements('id');
+        if (!Schema::connection($this->connection)->hasTable('logs')) {
+            Schema::connection($this->connection)->create('logs', function (Blueprint $table) {
+                $table->bigIncrements('id');
 
-            $table->uuid('session_id')->index();
-            $table->bigInteger('domain_id')->unsigned()->nullable()->index();
-            $table->bigInteger('uri_id')->unsigned()->index();
-            $table->string('event', 50)->nullable();
-            $table->string('entity_type', 60)->nullable();
-            $table->string('entity_id', 36)->nullable();
+                $table->uuid('session_id')->index();
+                $table->bigInteger('domain_id')->unsigned()->nullable()->index();
+                $table->bigInteger('uri_id')->unsigned()->index();
+                $table->string('event', 50)->nullable();
+                $table->string('entity_type', 60)->nullable();
+                $table->string('entity_id', 36)->nullable();
 
-            $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
+                $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
 
-            $table->foreign('session_id')->references('id')->on('sessions')->onDelete('cascade');
-            $table->foreign('domain_id')->references('id')->on('domains');
-            $table->foreign('uri_id')->references('id')->on('uris');
-        });
+                $table->foreign('session_id')->references('id')->on('sessions')->onDelete('cascade');
+                $table->foreign('domain_id')->references('id')->on('domains');
+                $table->foreign('uri_id')->references('id')->on('uris');
+            });
+        }
     }
 
     /**
